+++ src/main/java/com/patrol/api/member/member/controller/ApiV1AdminController.java
@@ -1,15 +1,14 @@
 package com.patrol.api.member.member.controller;
 
 import com.patrol.api.member.member.dto.GetAllMembersResponse;
-import com.patrol.domain.member.member.entity.Member;
+import com.patrol.api.member.member.dto.request.ChangeMemberStatusRequest;
+import com.patrol.domain.member.auth.service.AdminService;
 import com.patrol.domain.member.member.service.V2MemberService;
 import com.patrol.global.globalDto.GlobalResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 /**
  * packageName    : com.patrol.api.member.member.controller
@@ -26,11 +25,19 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RequestMapping("/api/v2/members/admin")
 public class ApiV1AdminController {
-    private final V2MemberService memberService;
+    private final AdminService adminService;
 
     // 관리자 > 회원 목록 조회
     @GetMapping
     public GlobalResponse<Page<GetAllMembersResponse>> getAllMembers(Pageable pageable) {
-        return GlobalResponse.success(memberService.getAllMembers(pageable));
+        return GlobalResponse.success(adminService.getAllMembers(pageable));
+    }
+
+    // 관리자 > 회원 상태 변경
+    @PatchMapping("/status")
+    public GlobalResponse<Void> changeMemberStatus(
+            @RequestBody ChangeMemberStatusRequest changeMemberStatusRequest) {
+        adminService.changeMemberStatus(changeMemberStatusRequest);
+        return GlobalResponse.success();
     }
 }
