--- src/main/java/com/patrol/global/security/RestTemplateConfig.java
@@ -1,25 +1,0 @@
-package com.patrol.global.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-/**
- * packageName    : com.patrol.global.security
- * fileName       : RestTemplatConfig
- * author         : sungjun
- * date           : 2025-03-04
- * description    : 자동 주석 생성
- * ===========================================================
- * DATE              AUTHOR             NOTE
- * -----------------------------------------------------------
- * 2025-03-04        kyd54       최초 생성
- */
-@Configuration
-public class RestTemplateConfig {
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-}
